Combine Salary and Other sections on PDS Support Items page
Merge the two breakdown tables into a single Support Items grid and
remove the formula math subtext to simplify the page (MPDX-9829). The
duplicate intermediate Salary Subtotal row is dropped since the combined
Subtotal already includes it.
