+++ src/components/HrTools/PdsGoalCalculator/SupportItem/SupportItemSectionList.test.tsx
@@ -3,7 +3,7 @@ import { PdsGoalCalculatorTestWrapper } from '../PdsGoalCalculatorTestWrapper';
 import { SupportItemSectionList } from './SupportItemSectionList';
 
 describe('SupportItemSectionList', () => {
-  it('renders Salary and Other as complete when setup is complete', async () => {
+  it('renders a single Support Items section as complete when setup is complete', async () => {
     const { findAllByRole } = render(
       <PdsGoalCalculatorTestWrapper>
         <SupportItemSectionList />
@@ -11,32 +11,29 @@ describe('SupportItemSectionList', () => {
     );
 
     const items = await findAllByRole('listitem');
-    expect(items).toHaveLength(2);
+    expect(items).toHaveLength(1);
 
-    const [salary, other] = items;
-    expect(salary).toHaveTextContent('Salary');
-    expect(other).toHaveTextContent('Other');
+    const [supportItems] = items;
+    expect(supportItems).toHaveTextContent('Support Items');
 
     await waitFor(() => {
-      expect(within(salary).getByTestId('CircleIcon')).toBeInTheDocument();
-      expect(within(other).getByTestId('CircleIcon')).toBeInTheDocument();
+      expect(
+        within(supportItems).getByTestId('CircleIcon'),
+      ).toBeInTheDocument();
     });
   });
 
-  it('renders Salary and Other as incomplete when setup is incomplete', async () => {
+  it('renders the Support Items section as incomplete when setup is incomplete', async () => {
     const { findAllByRole } = render(
       <PdsGoalCalculatorTestWrapper calculationMock={{ name: '' }}>
         <SupportItemSectionList />
       </PdsGoalCalculatorTestWrapper>,
     );
 
-    const [salary, other] = await findAllByRole('listitem');
+    const [supportItems] = await findAllByRole('listitem');
     await waitFor(() => {
       expect(
-        within(salary).getByTestId('RadioButtonUncheckedIcon'),
-      ).toBeInTheDocument();
-      expect(
-        within(other).getByTestId('RadioButtonUncheckedIcon'),
+        within(supportItems).getByTestId('RadioButtonUncheckedIcon'),
       ).toBeInTheDocument();
     });
   });
