+++ src/components/HrTools/PdsGoalCalculator/SupportItem/SupportItemStep.tsx
@@ -1,11 +1,38 @@
-import React from 'react';
+import React, { useMemo } from 'react';
 import { Box, Divider, Typography } from '@mui/material';
 import { useTranslation } from 'react-i18next';
-import { OtherSection } from './OtherSection';
-import { SalarySection } from './SalarySection';
+import { useLocale } from 'src/hooks/useLocale';
+import { useDataGridLocaleText } from 'src/hooks/useMuiLocaleText';
+import { usePdsGoalCalculator } from '../Shared/PdsGoalCalculatorContext';
+import { GridContainer, StyledGrid } from './styledGrid';
+import {
+  buildSupportItemBreakdownColumns,
+  buildSupportItemBreakdownRows,
+} from './supportItemBreakdown';
 
 export const SupportItemStep: React.FC = () => {
   const { t } = useTranslation();
+  const locale = useLocale();
+  const localeText = useDataGridLocaleText();
+  const { calculation, summaryData } = usePdsGoalCalculator();
+
+  const rows = useMemo(() => {
+    if (!calculation || !summaryData) {
+      return [];
+    }
+
+    return buildSupportItemBreakdownRows(
+      calculation,
+      summaryData.salaryConstants,
+      summaryData.otherConstants,
+      t,
+    );
+  }, [calculation, summaryData, t]);
+
+  const columns = useMemo(
+    () => buildSupportItemBreakdownColumns(locale, t),
+    [locale, t],
+  );
 
   return (
     <>
@@ -20,9 +47,34 @@ export const SupportItemStep: React.FC = () => {
         </Typography>
       </Box>
       <Divider sx={{ mx: -4, my: 4 }} />
-      <SalarySection />
-      <OtherSection />
-      <Divider sx={{ mx: -4, my: 4 }} />
+      {rows.length > 0 && (
+        <>
+          <GridContainer>
+            <StyledGrid
+              aria-label={t('Support items breakdown')}
+              rows={rows}
+              columns={columns}
+              getRowClassName={(params) => {
+                const classes: string[] = [];
+                if (params.id === 'subtotal') {
+                  classes.push('top-border');
+                  classes.push('bottom-border');
+                }
+                if (params.row.bold) {
+                  classes.push('bold-row');
+                }
+                return classes.join(' ');
+              }}
+              disableColumnMenu
+              disableColumnSorting
+              disableRowSelectionOnClick
+              hideFooter
+              localeText={localeText}
+            />
+          </GridContainer>
+          <Divider sx={{ mx: -4, my: 4 }} />
+        </>
+      )}
     </>
   );
 };
